Rename SSTChangePoint -> SingularSpectrumTransform
